+++ src/Codeception/Template/ModuleAcceptance/params.php
@@ -1,19 +1,24 @@
-<?php declare(strict_types=1);
+<?php
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
 
+declare(strict_types=1);
+
+use OxidEsales\Facts\Config\ConfigFile;
 use OxidEsales\Facts\Facts;
-use OxidEsales\Eshop\Core\ConfigFile;
-use OxidEsales\TestingLibrary\Services\Library\DatabaseDefaultsFileGenerator;
+use OxidEsales\Codeception\Module\Database\DatabaseDefaultsFileGenerator;
 
 $facts = new Facts();
+
 $selenium_server_port = getenv('SELENIUM_SERVER_PORT');
 $selenium_server_port = ($selenium_server_port) ? : '4444';
+$selenium_server_host = getenv('SELENIUM_SERVER_HOST');
+$selenium_server_host = ($selenium_server_host) ? : '127.0.0.1';
 $php = (getenv('PHPBIN')) ? : 'php';
-$cc_screen_shot_path = getenv('CC_SCREEN_SHOTS_PATH');
-$cc_screen_shot_path = ($cc_screen_shot_path) ? : '';
+$cc_screen_shot_url = getenv('CC_SCREEN_SHOTS_URL');
+$cc_screen_shot_url = ($cc_screen_shot_url) ? : '';
 
 return [
     'SHOP_URL' => $facts->getShopUrl(),
@@ -25,15 +30,22 @@ return [
     'DB_HOST' => $facts->getDatabaseHost(),
     'DB_PORT' => $facts->getDatabasePort(),
     'DUMP_PATH' => getTestDataDumpFilePath(),
+    'FIXTURES_PATH' => getTestFixtureSqlFilePath(),
     'MYSQL_CONFIG_PATH' => getMysqlConfigPath(),
     'SELENIUM_SERVER_PORT' => $selenium_server_port,
+    'SELENIUM_SERVER_HOST' => $selenium_server_host,
     'PHP_BIN' => $php,
-    'SCREEN_SHOT_URL' => $cc_screen_shot_path
+    'SCREEN_SHOT_URL' => $cc_screen_shot_url
 ];
 
 function getTestDataDumpFilePath()
 {
-    return getShopTestPath() . '/{{testsDir}}/{{dataDir}}/dump.sql';
+    return getShopTestPath() . '/Codeception/_data/generated/shop-dump.sql';
+}
+
+function getTestFixtureSqlFilePath()
+{
+    return getShopTestPath() . '/Codeception/_data/fixtures.sql';
 }
 
 function getShopSuitePath($facts)
@@ -42,7 +54,6 @@ function getShopSuitePath($facts)
     if (!$testSuitePath) {
         $testSuitePath = $facts->getShopRootPath() . '/tests';
     }
-
     return $testSuitePath;
 }
 
@@ -55,14 +66,12 @@ function getShopTestPath()
     } else {
         $shopTestPath = getShopSuitePath($facts);
     }
-
     return $shopTestPath;
 }
 
 function getMysqlConfigPath()
 {
-    $facts = new Facts();
-    $configFile = new ConfigFile($facts->getSourcePath() . '/config.inc.php');
+    $configFile = new ConfigFile();
 
     $generator = new DatabaseDefaultsFileGenerator($configFile);