+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.6.0] - unreleased
+
+### Added
+- ShopSetup module to setup test shop for tests
+
 ## [1.5.0] - 2021-03-25
 
 ### Added